Sonar round two: the CI scanner's deeper findings
The first token-authenticated Actions run swapped SonarCloud's automatic
analysis for the real Maven scanner - test sources and bytecode-aware
rules included - so the count ROSE after the sweep (114 -> 167) while
nothing regressed: the lens widened.

Fixed: all four S3516 blockers (multi-identical-return commands collapsed
to one return); nine provably-dead conditions (null checks against
@NonNull BentoBox/Paper APIs); a real rare crash - Math.abs(hashCode())%n
is negative at Integer.MIN_VALUE, now floorMod, in the resale-shelf pick;
setArmorContents(null) -> new ItemStack[4]; the customs Optional carried
as a value past its null gate; ~20 test classes of assertion/import/mock
hygiene; and the S3878 array-creation sites - dropped where the call has
one overload, replaced with a shared NO_VARS constant where
getTranslationAsComponent needs the disambiguation.

One lesson banked: a "provably dead" null check in boatId was shielding
bare Mockito mocks, not production - the check stayed deleted and the
mock now stubs a PersistentDataContainer, because production code is
entitled to lean on the API contract.

Remaining opens are deliberate and listed in docs/PROGRESS.md for
won't-fix marking: playSound's nullable-annotated-but-never-null Sound
constants, getLocation()-as-argument artifacts, deprecated event ctors in
tests, MapPalette.matchColor.

+++ src/main/java/world/bentobox/tradewinds/galaxy/GalaxyEngine.java
@@ -209,9 +209,9 @@ public class GalaxyEngine {
     private long computeBoostedTechCell() {
         long best = Long.MIN_VALUE;
         int bestRoll = -1;
-        for (long key : starterCells.stream().sorted().toList()) {
+        for (Long key : starterCells.stream().sorted().toList()) {
             int cellX = (int) (key >> 32);
-            int cellZ = (int) key;
+            int cellZ = key.intValue();
             IslandType type = cellX == 0 && cellZ == 0 && config.spawnIslandType() != null
                     ? config.spawnIslandType()
                     : rollType(cellX, cellZ);
@@ -338,7 +338,7 @@ public class GalaxyEngine {
     private int rollTech(int cellX, int cellZ, IslandType type) {
         int base = TECH_BASE.getOrDefault(type, 3);
         int wobble = Math.floorMod(Hashing.cellHash(config.seed(), cellX, cellZ, SALT_TECH), 5) - 2;
-        return Math.clamp(base + wobble, 1, MAX_TECH_LEVEL);
+        return Math.clamp((long) base + wobble, 1, MAX_TECH_LEVEL);
     }
 
     /**
